Remove line of code that can't be reached.
This line won't be reached ever, since we'll only ever go from state
index == boundary_length to index == boundary_length + 1 in the above
elif statement, and that statement will clear index if the value isn't a
hyphen or newline.  If it is one of the above, then one of our flags
will be set, and so we'll always have flags set in this condition.
Thus, the removed line should never be reached.