Add parsing for query string in request-body parser
This commit adds a method that will parse the query string given into
'request.GET', and moves the old form-body parsing to 'request.POST'.
Furthermore, the old 'request.fields' attribute is now a NestedMultiDict
that unifies the GET and POST attributes into a single dict.  Tests were
added for all of the above.
Also, some tests were modified for NestedMultiDict to deal with ordering
issues on Python 3.3.